@ Prepare 1.0.0 release: DBC float fix, choice numerics, changelog, sample
Pre-existing release-prep work (documented in CHANGELOG.md):
- DBC FLOAT attribute ranges parsed directly as double, fixing OverflowException
  on full single-precision scientific-notation bounds (e.g. -3.4E+038 3.4E+038).
- SignalValue numeric access (ToDouble/ToInt64/ToUInt64) returns the raw mapped
  value for decoded choices instead of throwing.
- Add CHANGELOG.md, version metadata, README notes, and samples/CanTools.Sample.

@

+++ src/CanTools/Formats/Dbc/DbcBuilder.cs
@@ -835,6 +835,10 @@ internal sealed class DbcBuilder
     private static long ToInt(string text) =>
         (long)decimal.Parse(text, NumberStyles.Float, CultureInfo.InvariantCulture);
 
+    // Upstream parses via Decimal purely to be liberal about the literal form;
+    // it then converts to a Python float (a double). Parse straight to double so
+    // the full single-precision range (e.g. FLOAT bounds like -3.4E+038) fits —
+    // System.Decimal's range is far smaller and would overflow on those.
     private static double ToFloat(string text) =>
-        (double)decimal.Parse(text, NumberStyles.Float, CultureInfo.InvariantCulture);
+        double.Parse(text, NumberStyles.Float, CultureInfo.InvariantCulture);
 }
